[NEW] Calculo de imposto do cabeçalho dos documentos
